docker-compose written com.mysql.cj.jdbc.exceptions.CommunicationsException: Communications link failure

+++ src/main/java/searchengine/dto/Lemma2PageDto.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Lemma2PageDto {
-    private int id;
-    private int pageId;
-    private int lemmaId;
-    private float quantity;
+    private Long id;
+    private Long pageId;
+    private Long lemmaId;
+    private Float quantity;
 }
