+++ src/main/java/searchengine/model/Lemma.java
@@ -1,6 +1,5 @@
 package searchengine.model;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -9,17 +8,13 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 
-import java.util.List;
 
 @Getter
 @Setter
@@ -29,10 +24,9 @@ import java.util.List;
 public class Lemma {
     @Id
     @Column(name = "id")
-//    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_generator")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @OnDelete(action = OnDeleteAction.CASCADE)
-    private int id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "site_id", referencedColumnName = "id", nullable = false)
@@ -42,10 +36,7 @@ public class Lemma {
     private String lemma;
 
     @Column(name = "frequency", nullable = false)
-    private int frequency;
-
-
-
+    private Long frequency;
 
 
     @Override
