+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -12,13 +12,16 @@ import java.util.List;
 @Repository
 public interface PageRepository extends JpaRepository<Page, Integer> {
     @Query(value = "SELECT count(*) FROM search_engine.page where site_id = ?1", nativeQuery = true)
-    int countIndexedPages(int id);
+    int countIndexedPages(Long id);
+
     @Query(value = "select * from search_engine.page where path = ?1", nativeQuery = true)
-    int getPageIdByUrl(String url);
+    Long getPageIdByUrl(String url);
+
     @Query(value = "select count(*) from search_engine.page", nativeQuery = true)
     int countPages();
+
     @Query(value = "select count(*) from search_engine.page where site_id = ?1", nativeQuery = true)
-    int countPagesBySiteId(int siteId);
+    int countPagesBySiteId(Long siteId);
 
     @Query(value = "select * from search_engine.page where path = ?1", nativeQuery = true)
     Page getPageByUrl(String url);
@@ -26,7 +29,7 @@ public interface PageRepository extends JpaRepository<Page, Integer> {
     @Transactional
     @Modifying
     @Query(value = "delete from search_engine.page where id = ?1", nativeQuery = true)
-    void deletePage(int id);
+    void deletePage(Long id);
 
     @Query(value = "select * from search_engine.page where id in (select page_id from search_engine.l2p where lemma_id = ?1)", nativeQuery = true)
     List<Page> getPageListByLemmaId(int lemmaId);
